requirements.txt
openai
python-dotenv
pywikibot
tiktoken